Updating the wheels-cli to run testbox tests.
Updated the wheels-cli so that when a user runs "wheels test" command, it will run the testbox tests instead of the rocketUnit tests. User can pass the "type" argument with a value of either "app" or "core". Passing the "app" value runs the testbox tests for the application while passing the "core" value runs the framework tests.