Upgrade the snippet copying process
Updated the code for the ensureSnippetTemplatesExist function. Now, it checks within the folders it is copying and copies the files from inside them that are missing from the application.
Added the function copyMissingFolderContents which handles the copying of files from inside the folders. It also handles any nested folders as well.