Update template copying porcess
Updated the template copying process so that it only copies the files if the app folder is present.
Also, now copying all the files instead of excluding the bootstrap folder and other files.
Also updated the getTemplateDirectory method to use the templates folder from the cli if the app/snippets folder is not present